Changes in Catalog APIs for better filtering in User Module

+++ services/app/core/database.py
@@ -69,40 +69,73 @@ async def retrieve_books():
 def format_catalog(catalog) -> dict:
     return {
         "catalog_name": catalog["name"],
+        "user": catalog["user_id"],
         "description": catalog["description"],
         "books": catalog["books_list"]
     }
 
 async def create_catalog(catalog_data: dict):
-    catalog = await catalog_collection.insert_one(
-        { "name": catalog_data.name, "description": catalog_data.description, "books_list": catalog_data.books_list}
-    )
-    print(catalog_data.name)
-    new_catalog = await catalog_collection.find_one({"name": catalog_data.name})
+    catalog_names = []
+    async for catalog in catalog_collection.find():
+        catalog_names.append(catalog["name"])
+    if catalog_data.name not in catalog_names:
+        catalog = await catalog_collection.insert_one(
+        { "name": catalog_data.name, "user_id": catalog_data.user_id, "description": catalog_data.description, "books_list": catalog_data.books_list})
+    new_catalog = await catalog_collection.find_one({"name": catalog_data.name, "user_id": catalog_data.user_id})
     return format_catalog(new_catalog)
 
-async def get_catalog(catalog_name: str) -> dict:
-    curr_catalog = await catalog_collection.find_one({"name": catalog_name})
+async def get_all_catalogs_of_user(user_id : str, catalog_name: str) -> dict:
+    curr_catalog = await catalog_collection.find_one({"name": catalog_name,"user_id": user_id })
     return format_catalog(curr_catalog)
 
-async def update_catalog_book_list(catalog_name:str, book_isbn_id: str) -> dict:
-    curr_catalog = await catalog_collection.find_one({"name": catalog_name})
+async def add_book_to_catalog(user_id : str, catalog_name:str, book_isbn_id: str) -> dict:
+    curr_catalog = await catalog_collection.find_one({"name": catalog_name,"user_id": user_id})
     if curr_catalog is not None:
         logger.info("Catalog Exists")
-        update_criteria = { "name": catalog_name }
+        update_criteria = { "name": catalog_name,"user_id": user_id }
         curr_catalog["books_list"].append(book_isbn_id)
         newvalues = {"$set": { "name": curr_catalog["name"] ,
-                               "description": curr_catalog["description"] ,
-                             "books_list": curr_catalog["books_list"]} }
+                                "user_id" : curr_catalog["user_id"],
+                                "description": curr_catalog["description"] ,
+                                "books_list": curr_catalog["books_list"]} }
 
     catalog_updated = await catalog_collection.update_one(update_criteria, newvalues)
     return catalog_updated.acknowledged
 
-async def delete_catalog(catalog_name: str):
-    catalog_available = await catalog_collection.find_one({"name": catalog_name})
+
+async def update_catalog_book_list(user_id : str, catalog_name:str, new_books_list: list) -> dict:
+    curr_catalog = await catalog_collection.find_one({"name": catalog_name,"user_id": user_id})
+    if curr_catalog is not None:
+        logger.info("Catalog Exists")
+        update_criteria = { "name": catalog_name,"user_id": user_id }
+        for book_isbn_id in new_books_list:
+            if book_isbn_id not in curr_catalog["books_list"]:
+                curr_catalog["books_list"].append(book_isbn_id)
+        newvalues = {"$set": { "name": curr_catalog["name"] ,
+                                "user_id" : curr_catalog["user_id"],
+                                "description": curr_catalog["description"] ,
+                                "books_list": curr_catalog["books_list"]} }
+    catalog_updated = await catalog_collection.update_one(update_criteria, newvalues)
+    return catalog_updated.acknowledged
+
+
+async def delete_catalog(user_id : str, catalog_name: str):
+    catalog_available = await catalog_collection.find_one({"name": catalog_name, "user_id": user_id})
     if catalog_available:
-        await catalog_collection.delete_one({"name": catalog_name})
+        await catalog_collection.delete_one({"name": catalog_name, "user_id": user_id})
     return True
+
+async def delete_books_from_catalog(user_id : str, catalog_name: str, books_remove_list: list):
+    catalog_available = await catalog_collection.find_one({"name": catalog_name, "user_id": user_id})
+    if catalog_available:
+        new_books_list = [ele for ele in catalog_available["books_list"] if ele not in books_remove_list]
+        update_criteria = { "name": catalog_name,"user_id": user_id }
+        newvalues = {"$set": { "name": catalog_available["name"] ,
+                                "user_id" : catalog_available["user_id"],
+                                "description": catalog_available["description"] ,
+                                "books_list": new_books_list} }
+    catalog_updated = await catalog_collection.update_one(update_criteria, newvalues)
+    return catalog_updated.acknowledged
 
 async def retrieve_catalogs():
     catalogs = []
